+++ rentals.json
@@ -1,0 +1,1 @@
+[{"date":"2022/1/1","book":{"title":"greg","author":"meme"},"person":{"name":"earl","age":"24","id":705,"type":"Teacher","specialization":"dec"}}]
